create a submodule for db_import
