+++ config.ini
@@ -7,6 +7,8 @@ chroma_host=localhost
 chroma_port=8000
 chroma_collection=buildragwithpython
 conversations_dir=./conversations
+
 [chat]
-promt_system="You are a helpful assistant that is an expert at extracting the most useful information from a given text."
+prompt_system="You are a helpful assistant that is an expert at extracting the most useful information from a given text."
+
 [import]
