+++ import.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
-from utilities import readtext, getconfig
-from mattsollamatools import chunker, chunk_text_by_sentences
+from utilities import getconfig
+from db_import.file_import import import_file
 import ollama, chromadb, time
 import sys
 import argparse
@@ -24,13 +24,7 @@ starttime = time.time()
 with sys.stdin as f:
   lines = f.readlines()
   for filename in lines:
-    text = readtext(filename)
-    chunks = chunk_text_by_sentences(source_text=text, sentences_per_chunk=7, overlap=0 )
-    print(f"with {len(chunks)} chunks")
-    for index, chunk in enumerate(chunks):
-      embed = ollama.embeddings(model=embedmodel, prompt=chunk)['embedding']
-      print(".", end="", flush=True)
-      collection.add([filename+str(index)], [embed], documents=[chunk], metadatas={"source": filename})
+    import_file(filename, embedmodel, collection)
     
 print("--- %s seconds ---" % (time.time() - starttime))
