+++ utilities.py
@@ -26,39 +26,10 @@ def download_file(url):
                 f.write(chunk)
         return filename
       
-def readtext(path):
-  path = path.rstrip()
-  path = path.replace(' \n', '')
-  path = path.replace('%0A', '')
-  if re.match(r'^https?://', path):
-    filename = download_file(path)
-  else:
-    
-    relative_path = path
-    filename = os.path.abspath(relative_path)
-  
-  filetype = magic.from_file(filename, mime=True)
-  print(f"\nEmbedding {filename} as {filetype}")
-  text = ""
-  if filetype == 'application/pdf':
-    print('PDF not supported yet')
-  if filetype == 'text/plain':
-    with open(filename, 'rb') as f:
-      text = f.read().decode('utf-8')
-  if filetype == 'text/html':
-    with open(filename, 'rb') as f:
-      soup = BeautifulSoup(f, 'html.parser')
-      text = soup.get_text()
-  
-  if os.path.exists(filename) and filename.find('content/') > -1:
-    os.remove(filename) 
-    
-  return text
-
 config = False
 def getconfig(section, key):
   global config
   if config == False:
     config = configparser.ConfigParser()
     config.read('config.ini')
-  return dict(config.items("main"))[key]
+  return dict(config.items(section))[key]
